UIIN-2943: Create separate TitleCell component
