+++ src/components/TitlesView/TitlesView.js
@@ -1,24 +1,15 @@
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { Link } from 'react-router-dom';
 
 import {
   MultiColumnList,
   KeyValue,
-  IconButton,
-  Tooltip,
+  NoValue,
 } from '@folio/stripes/components';
-import {
-  queryIndexes,
-  ADVANCED_SEARCH_INDEX,
-  advancedSearchQueryBuilder,
-} from '@folio/stripes-inventory-components';
 
-import {
-  getIdentifiers,
-  getIdentifiersValues,
-} from '../../utils';
+import { getIdentifiers } from '../../utils';
 import { indentifierTypeNames } from '../../constants';
+import TitleCell from './TitleCell';
 
 const TitlesViews = ({
   titles,
@@ -31,74 +22,18 @@ const TitlesViews = ({
     ISSN,
     ISBN,
   } = indentifierTypeNames;
-  const createInstanceLink = (row) => {
-    const identifiersValues = getIdentifiersValues(row.identifiers, [ISBN, ISSN], identifierTypesById);
-    const isbn = identifiersValues?.[ISBN];
-    const issn = identifiersValues?.[ISSN];
-
-    const advancedSearchRows = [
-      {
-        bool: '',
-        match: 'exactPhrase',
-        searchOption: queryIndexes.TITLE,
-        query: row.title,
-      },
-      ...(isbn ? [{
-        bool: 'and',
-        match: 'containsAll',
-        searchOption: queryIndexes.ISBN,
-        query: isbn,
-      }] : []),
-      ...(issn ? [{
-        bool: 'and',
-        match: 'containsAll',
-        searchOption: queryIndexes.ISSN,
-        query: issn,
-      }] : [])
-    ];
-
-    const builtQuery = advancedSearchQueryBuilder(advancedSearchRows);
-
-    const searchParams = new URLSearchParams({
-      qindex: ADVANCED_SEARCH_INDEX,
-      query: builtQuery,
-    });
-
-    return `/inventory?${searchParams.toString()}`;
-  };
 
   const formatter = {
-    title: row => (row[titleKey]
-      ?
-        <Link
-          to={`/inventory/view/${row[titleKey]}`}
-        >
-          {row.title}
-        </Link>
-      : row.title
-        ?
-          <>
-            {row.title}
-            <Tooltip
-              id="searchForTitle"
-              text={<FormattedMessage id="ui-inventory.searchForTitle" values={{ title: row.title }} />}
-            >
-              {({ ref, ariaIds }) => (
-                <IconButton
-                  ref={ref}
-                  icon="search"
-                  to={createInstanceLink(row)}
-                  target="_blank"
-                  aria-labelledby={ariaIds.text}
-                />
-              )}
-            </Tooltip>
-          </>
-        : '-'
+    title: row => (
+      <TitleCell
+        rowData={row}
+        titleKey={titleKey}
+        identifierTypesById={identifierTypesById}
+      />
     ),
-    hrid: row => row.hrid || '-',
-    issn: row => getIdentifiers(row.identifiers, ISSN, identifierTypesById) || '-',
-    isbn: row => getIdentifiers(row.identifiers, ISBN, identifierTypesById) || '-',
+    hrid: row => row.hrid || <NoValue />,
+    issn: row => getIdentifiers(row.identifiers, ISSN, identifierTypesById) || <NoValue />,
+    isbn: row => getIdentifiers(row.identifiers, ISBN, identifierTypesById) || <NoValue />,
   };
 
   const visibleColumns = [
